+++ NEWS.md
@@ -119,3 +119,6 @@
 * Fix issue showing individual bootstrap points in 3D
 * Add option `show_group_legend` to `view()` to show interactive legend based on `agGroups()` and `srGroups()` point groupings
 * Add ability in viewer to filter and color antigens and sera by group
+
+# Racmacs 1.1.28
+* `agStress()` and `srStress()` fixed to work with NA coordinates